requirements.txt
pyyaml
Flask
flask-htpasswd
python-telegram-bot